move GrpcConfig
