+++ src/SkyApm.Transport.Grpc/V8/LoggerReporter.cs
@@ -18,6 +18,7 @@
 
 using SkyApm.Config;
 using SkyApm.Logging;
+using SkyApm.Transport.Grpc.Common;
 using SkyWalking.NetworkProtocol.V3;
 using System;
 using System.Collections.Generic;
